requirements.txt
Flask
openai==0.28.0
gunicorn
python-dotenv
